rotations rearrange
